Checkbox moved to molecule folder, style compoenents relocated and imported, all imports amended accordingly
