+++ src/components/styles/molecules/CheckboxStyling.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledCheckboxField = styled.input`
+width: 20px;
+height: 20px;
+margin: 0 20px 10px 0;
+`;
